+++ clevertap-core/src/main/java/com/clevertap/android/sdk/CleverTapInstanceConfig.java
@@ -57,11 +57,7 @@ public class CleverTapInstanceConfig implements Parcelable {
 
     private boolean disableAppLaunchedEvent;
 
-    private boolean enableABTesting;
-
     private boolean enableCustomCleverTapId;
-
-    private boolean enableUIEditor;
 
     private String fcmSenderId;
 
@@ -118,8 +114,6 @@ public class CleverTapInstanceConfig implements Parcelable {
         this.backgroundSync = config.backgroundSync;
         this.enableCustomCleverTapId = config.enableCustomCleverTapId;
         this.fcmSenderId = config.fcmSenderId;
-        this.enableABTesting = config.enableABTesting;
-        this.enableUIEditor = config.enableUIEditor;
         this.packageName = config.packageName;
         this.beta = config.beta;
         this.allowedPushTypes = config.allowedPushTypes;
@@ -138,8 +132,6 @@ public class CleverTapInstanceConfig implements Parcelable {
         this.debugLevel = CleverTapAPI.LogLevel.INFO.intValue();
         this.logger = new Logger(this.debugLevel);
         this.createdPostAppLaunch = false;
-        this.enableABTesting = this.isDefaultInstance;
-        this.enableUIEditor = this.enableABTesting;
 
         ManifestInfo manifest = ManifestInfo.getInstance(context);
         this.useGoogleAdId = manifest.useGoogleAdId();
@@ -192,12 +184,7 @@ public class CleverTapInstanceConfig implements Parcelable {
                 this.debugLevel = configJsonObject.getInt(Constants.KEY_DEBUG_LEVEL);
             }
             this.logger = new Logger(this.debugLevel);
-            if (configJsonObject.has(Constants.KEY_ENABLE_ABTEST)) {
-                this.enableABTesting = configJsonObject.getBoolean(Constants.KEY_ENABLE_ABTEST);
-            }
-            if (configJsonObject.has(Constants.KEY_ENABLE_UIEDITOR)) {
-                this.enableUIEditor = configJsonObject.getBoolean(Constants.KEY_ENABLE_UIEDITOR);
-            }
+
             if (configJsonObject.has(Constants.KEY_PACKAGE_NAME)) {
                 this.packageName = configJsonObject.getString(Constants.KEY_PACKAGE_NAME);
             }
@@ -250,8 +237,6 @@ public class CleverTapInstanceConfig implements Parcelable {
         backgroundSync = in.readByte() != 0x00;
         enableCustomCleverTapId = in.readByte() != 0x00;
         fcmSenderId = in.readString();
-        enableABTesting = in.readByte() != 0x00;
-        enableUIEditor = in.readByte() != 0x00;
         packageName = in.readString();
         logger = new Logger(debugLevel);
         beta = in.readByte() != 0x00;
@@ -324,11 +309,6 @@ public class CleverTapInstanceConfig implements Parcelable {
         return identityKeys;
     }
 
-    @SuppressWarnings({"BooleanMethodIsAlwaysInverted", "WeakerAccess"})
-    public boolean isABTestingEnabled() {
-        return enableABTesting;
-    }
-
     @SuppressWarnings({"unused", "WeakerAccess"})
     public boolean isAnalyticsOnly() {
         return analyticsOnly;
@@ -347,11 +327,6 @@ public class CleverTapInstanceConfig implements Parcelable {
         return isDefaultInstance;
     }
 
-    @SuppressWarnings({"unused"})
-    public boolean isUIEditorEnabled() {
-        return enableUIEditor;
-    }
-
     @RestrictTo(RestrictTo.Scope.LIBRARY)
     public void log(@NonNull String tag, @NonNull String message) {
         logger.verbose(getDefaultSuffix(tag), message);
@@ -360,18 +335,6 @@ public class CleverTapInstanceConfig implements Parcelable {
     @RestrictTo(RestrictTo.Scope.LIBRARY)
     public void log(@NonNull String tag, @NonNull String message, Throwable throwable) {
         logger.verbose(getDefaultSuffix(tag), message, throwable);
-    }
-
-    @SuppressWarnings("SameParameterValue")
-    @RestrictTo(RestrictTo.Scope.LIBRARY)
-    public void setEnableABTesting(boolean enableABTesting) {
-        this.enableABTesting = enableABTesting;
-    }
-
-    @SuppressWarnings({"unused"})
-    @RestrictTo(RestrictTo.Scope.LIBRARY)
-    public void setEnableUIEditor(boolean enableUIEditor) {
-        this.enableUIEditor = enableUIEditor;
     }
 
     public void setIdentityKeys(@IdentityType String... identityKeys) {
@@ -403,8 +366,6 @@ public class CleverTapInstanceConfig implements Parcelable {
         dest.writeByte((byte) (backgroundSync ? 0x01 : 0x00));
         dest.writeByte((byte) (enableCustomCleverTapId ? 0x01 : 0x00));
         dest.writeString(fcmSenderId);
-        dest.writeByte((byte) (enableABTesting ? 0x01 : 0x00));
-        dest.writeByte((byte) (enableUIEditor ? 0x01 : 0x00));
         dest.writeString(packageName);
         dest.writeByte((byte) (beta ? 0x01 : 0x00));
         dest.writeList(allowedPushTypes);
@@ -487,8 +448,6 @@ public class CleverTapInstanceConfig implements Parcelable {
             configJsonObject.put(Constants.KEY_ENABLE_CUSTOM_CT_ID, getEnableCustomCleverTapId());
             configJsonObject.put(Constants.KEY_PACKAGE_NAME, getPackageName());
             configJsonObject.put(Constants.KEY_BETA, isBeta());
-            configJsonObject.put(Constants.KEY_ENABLE_UIEDITOR, isUIEditorEnabled());
-            configJsonObject.put(Constants.KEY_ENABLE_ABTEST, isABTestingEnabled());
             configJsonObject.put(Constants.KEY_ALLOWED_PUSH_TYPES, toJsonArray(allowedPushTypes));
             configJsonObject.put(Constants.KEY_STAGING, getStaging());
             return configJsonObject.toString();