+++ clevertap-core/src/main/java/com/clevertap/android/sdk/ManifestInfo.java
@@ -69,33 +69,33 @@ public class ManifestInfo {
             metaData = new Bundle();
         }
         if (accountId == null) {
-            accountId = _getManifestStringValueForKey(metaData, Constants.LABEL_ACCOUNT_ID);
+            accountId = metaData.getString(Constants.LABEL_ACCOUNT_ID);
         }
         if (accountToken == null) {
-            accountToken = _getManifestStringValueForKey(metaData, Constants.LABEL_TOKEN);
+            accountToken = metaData.getString(Constants.LABEL_TOKEN);
         }
         if (accountRegion == null) {
-            accountRegion = _getManifestStringValueForKey(metaData, Constants.LABEL_REGION);
+            accountRegion = metaData.getString(Constants.LABEL_REGION);
         }
-        notificationIcon = _getManifestStringValueForKey(metaData, Constants.LABEL_NOTIFICATION_ICON);
-        useADID = "1".equals(_getManifestStringValueForKey(metaData, Constants.LABEL_USE_GOOGLE_AD_ID));
-        appLaunchedDisabled = "1".equals(_getManifestStringValueForKey(metaData, Constants.LABEL_DISABLE_APP_LAUNCH));
-        excludedActivities = _getManifestStringValueForKey(metaData, Constants.LABEL_INAPP_EXCLUDE);
-        sslPinning = "1".equals(_getManifestStringValueForKey(metaData, Constants.LABEL_SSL_PINNING));
-        backgroundSync = "1".equals(_getManifestStringValueForKey(metaData, Constants.LABEL_BACKGROUND_SYNC));
-        useCustomID = "1".equals(_getManifestStringValueForKey(metaData, Constants.LABEL_CUSTOM_ID));
-        fcmSenderId = _getManifestStringValueForKey(metaData, Constants.LABEL_FCM_SENDER_ID);
+        notificationIcon = metaData.getString(Constants.LABEL_NOTIFICATION_ICON);
+        useADID = "1".equals(metaData.getString(Constants.LABEL_USE_GOOGLE_AD_ID));
+        appLaunchedDisabled = "1".equals(metaData.getString(Constants.LABEL_DISABLE_APP_LAUNCH));
+        excludedActivities = metaData.getString(Constants.LABEL_INAPP_EXCLUDE);
+        sslPinning = "1".equals(metaData.getString(Constants.LABEL_SSL_PINNING));
+        backgroundSync = "1".equals(metaData.getString(Constants.LABEL_BACKGROUND_SYNC));
+        useCustomID = "1".equals(metaData.getString(Constants.LABEL_CUSTOM_ID));
+        fcmSenderId = metaData.getString(Constants.LABEL_FCM_SENDER_ID);
         if (fcmSenderId != null) {
             fcmSenderId = fcmSenderId.replace("id:", "");
         }
-        packageName = _getManifestStringValueForKey(metaData, Constants.LABEL_PACKAGE_NAME);
-        beta = "1".equals(_getManifestStringValueForKey(metaData, Constants.LABEL_BETA));
+        packageName = metaData.getString(Constants.LABEL_PACKAGE_NAME);
+        beta = "1".equals(metaData.getString(Constants.LABEL_BETA));
         if (intentServiceName == null) {
-            intentServiceName = _getManifestStringValueForKey(metaData, Constants.LABEL_INTENT_SERVICE);
+            intentServiceName = metaData.getString(Constants.LABEL_INTENT_SERVICE);
         }
 
-        xiaomiAppKey = _getManifestStringValueForKey(metaData, Constants.LABEL_XIAOMI_APP_KEY);
-        xiaomiAppID = _getManifestStringValueForKey(metaData, Constants.LABEL_XIAOMI_APP_ID);
+        xiaomiAppKey = metaData.getString(Constants.LABEL_XIAOMI_APP_KEY);
+        xiaomiAppID = metaData.getString(Constants.LABEL_XIAOMI_APP_ID);
 
         if (LC == null) {
             LC = _getManifestStringValueForKey(metaData, Constants.LABEL_LC);
@@ -177,7 +177,7 @@ public class ManifestInfo {
     }
 
     private String[] parseProfileKeys(final Bundle metaData) {
-        String profileKeyString = _getManifestStringValueForKey(metaData, Constants.CLEVERTAP_IDENTIFIER);
+        String profileKeyString = metaData.getString(Constants.CLEVERTAP_IDENTIFIER);
         return !TextUtils.isEmpty(profileKeyString) ? profileKeyString.split(Constants.SEPARATOR_COMMA)
                 : NullConstants.NULL_STRING_ARRAY;
     }
@@ -188,12 +188,4 @@ public class ManifestInfo {
         accountRegion = region;
     }
 
-    private static String _getManifestStringValueForKey(Bundle manifest, String name) {
-        try {
-            Object o = manifest.get(name);
-            return (o != null) ? o.toString() : null;
-        } catch (Throwable t) {
-            return null;
-        }
-    }
 }