+++ clevertap-core/src/main/java/com/clevertap/android/sdk/task/MainThreadExecutor.java
@@ -7,7 +7,11 @@ import java.util.concurrent.Executor;
 
 public class MainThreadExecutor implements Executor {
 
-    private final Handler mainThreadHandler = new Handler(Looper.getMainLooper());
+    void setMainThreadHandler(final Handler mainThreadHandler) {
+        this.mainThreadHandler = mainThreadHandler;
+    }
+
+    Handler mainThreadHandler = new Handler(Looper.getMainLooper());
 
     @Override
     public void execute(@NonNull Runnable command) {